+++ src/app/models/key-models/listlikekey.cpp
@@ -1,7 +1,7 @@
 #include "listlikekey.h"
 
 ListLikeKeyModel::ListLikeKeyModel(QSharedPointer<RedisClient::Connection> connection,
-                                   QString fullPath, int dbIndex, int ttl,
+                                   QString fullPath, int dbIndex, long long ttl,
                                    QByteArray rowsCountCmd,
                                    QByteArray partialLoadingCmd, QByteArray fullLoadingCmd,
                                    bool fullLoadingCmdSupportsRanges)
